requirements.txt
# Data manipulation
pandas>=2.0.0
numpy>=1.24.0

# Web scraping and API requests
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
python-dotenv>=1.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Machine Learning
scikit-learn>=1.3.0
xgboost>=2.0.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.25.0

# Optional: for Streamlit app (Step 7.3)
streamlit>=1.28.0
